requirements.txt
Django==2.2.9
django-appconf==1.0.3
django-imagekit==4.0.2
djangorestframework==3.11.0
pilkit==2.0
Pillow==6.2.1
psycopg2==2.7.4
psycopg2-binary==2.8.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
